requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.3.0
distlib==0.3.7
Django==4.2.5
djangorestframework==3.14.0
drf-spectacular==0.26.5
filelock==3.12.4
frozenlist==1.4.0
gunicorn==21.2.0
identify==2.5.30
idna==3.4
inflection==0.5.1
jsonschema==3.2.0
multidict==6.0.4
nodeenv==1.8.0
openai==0.28.1
packaging==23.2
platformdirs==3.11.0
pre-commit==3.4.0
psycopg2-binary==2.9.8
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
uritemplate==4.1.1
urllib3==2.0.6
virtualenv==20.24.5
yarl==1.9.2
